pyproject.toml
[project]
name = "okx-hbot-ai-agent-mvp"
version = "0.1.0"
description = "MVP scaffold for OKX + Hummingbot + AI Agent trading copilot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "python-okx>=0.4.1"
]

[tool.unittest]
start-directory = "tests"
